requirements.txt
ccxt
pandas
numpy
pandas-ta
requests
python-dotenv

